+++ server/src/modules/admin/adminActions.ts
@@ -1,7 +1,8 @@
 import type { RequestHandler } from "express";
 
 // Import access to data
-import adminRepository, { type User } from "./adminRepository";
+import type { User } from "../../types/admin/adminTypes";
+import adminRepository from "./adminRepository";
 
 // The B of BREAD - Browse (Read All) operation
 const browse: RequestHandler = async (req, res, next) => {
