+++ server/src/types/admin/adminTypes.ts
@@ -1,0 +1,12 @@
+export type Admin = {
+  id: number;
+  user_id: number;
+};
+
+export type User = {
+  id: number;
+  email: string;
+  password: string;
+  firstname: string;
+  lastname: string;
+};
